test(test): split files and add more comprehensive tests
